Safari-609.1.11.1/LayoutTests/inspector/dom/shadow-and-non-shadow-children-expected.txt

Test for DOMNode.children of an element with a shadow and non-shadow children.



== Running test suite: DOMNode.ShadowAndNonShadowChildren
-- Running test case: DOMNode.ShadowAndNonShadowChildren.VideoWithSource
PASS: DOMNode should be a `video` element.
PASS: DOMNode should have child nodes.
PASS: DOMNode should indicate it has 2 child nodes.
PASS: DOMNode children should be null.
PASS: DOMNode should have 2 children.
PASS: DOMNode 1st child should be a ShadowRoot.
PASS: DOMNode 2nd child should be a `source` element.
